fix round data not read from the cache
